Ensure the downloaded file gets named as listed in the feed
Feeds often redirect to a different URL to download the file rather than
the one listed in the feed.  Most of the time, the filename at the
redirected URL is the same as the one listed in the feed.  However,
sometimes it is different (eg http://rss.cat5.tv/hd.rss).  In the case
of this feed before this change, the file would be named a 'random' nine
digit number (a vimeo ID I think as it comes from vimeocdn).  The feed
has a well named filename.  This change takes the well named filename
and uses it for the name of the saved file.  This also uses the extension
of the download location rather than the feed name.  Again, in the example
above, the feed lists the extension as m4v, and the downloaded file is
a mp4.  While having the downloaded file listed as m4v when it is actually
a mp4 may work without issue, it just makes more sense to use the actual
extension of the downloaded file, which this commit does.